Rewrote function dst() to depend on pytz rather than time
When we were calling time.gmtime() to get a tm struct, it was never going
to set the dst flag since UTC doesn't define DST, so some of the tests
could never work. This update looks through pytz._utc_transition_times[] to
find the time segment containing dst()'s *when* argument, then looks at
pytz._transition_info[] to decide whether DST is in effect or not by
considering whether the DST offset is 0 (off) or not (on).